PKG-872 Packaging tasks for release - PS 8.4.6-6

+++ storage/innobase/btr/btr0sea.cc
@@ -419,8 +419,7 @@ static void btr_search_info_update_hash(btr_cur_t *cursor) {
     return;
   }
 
-  const uint16_t n_unique =
-      static_cast<uint16_t>(dict_index_get_n_unique_in_tree(index));
+  const uint16_t n_unique = dict_index_get_n_unique_in_tree(index);
   const auto info = index->search_info;
   if (info->n_hash_potential != 0) {
     const auto prefix_info = info->prefix_info.load();
